refactor(runtime): extract process / canvas / skill / artifact tool modules
Four more leaf modules pulled out of tool_runner/mod.rs:

- `process` — start / poll / write / kill / list, each a thin wrapper
  over crate::process_manager::ProcessManager.
- `canvas` — sanitize_canvas_html + tool_canvas_present. The sanitizer
  keeps its `pub fn` visibility and is re-exported from tool_runner
  via `pub use self::canvas::sanitize_canvas_html`, so any
  `librefang_runtime::tool_runner::sanitize_canvas_html` consumer
  continues to resolve. canvas_present pulls CANVAS_MAX_BYTES from
  the parent module's task-local declaration.
- `skill` — the eight skill-evolution tools (create / update / patch /
  delete / rollback / write_file / remove_file) plus tool_skill_read_file,
  with the agent_author_tag and ensure_not_frozen helpers kept
  module-private since they are only consumed by this module.
- `artifact` — tool_read_artifact (#3347).

Verbatim function moves; only visibility upgraded to `pub(super)` (or
`pub` where the symbol was already crate-public). No behavior change.

mod.rs: 12,360 -> 11,744 LOC (-616). cargo check + cargo clippy --lib
both clean. The orphan `// ---` separator that the previous Docker /
process boundary left behind has also been removed.

Refs #3710
