+++ crates/librefang-runtime/src/tool_runner/mod.rs
@@ -16,6 +16,8 @@ use tracing::{debug, warn};
 
 mod a2a;
 mod agent;
+mod artifact;
+mod canvas;
 mod cron;
 mod event;
 mod goal;
@@ -23,8 +25,10 @@ mod hand;
 mod knowledge;
 mod memory;
 mod notify;
+mod process;
 mod schedule;
 mod shell_safety;
+mod skill;
 mod system;
 mod taint;
 mod task;
@@ -33,6 +37,9 @@ mod workflow;
 
 use self::a2a::{tool_a2a_discover, tool_a2a_send};
 use self::agent::tool_agent_find;
+use self::artifact::tool_read_artifact;
+pub use self::canvas::sanitize_canvas_html;
+use self::canvas::tool_canvas_present;
 use self::cron::{tool_cron_cancel, tool_cron_create, tool_cron_list};
 use self::event::tool_event_publish;
 use self::goal::tool_goal_update;
@@ -42,8 +49,16 @@ use self::knowledge::{
 };
 use self::memory::{tool_memory_list, tool_memory_recall, tool_memory_store};
 use self::notify::tool_notify_owner;
+use self::process::{
+    tool_process_kill, tool_process_list, tool_process_poll, tool_process_start, tool_process_write,
+};
 use self::schedule::{tool_schedule_create, tool_schedule_delete, tool_schedule_list};
 use self::shell_safety::{classify_shell_exec_ro_safety, RoSafety};
+use self::skill::{
+    tool_skill_evolve_create, tool_skill_evolve_delete, tool_skill_evolve_patch,
+    tool_skill_evolve_remove_file, tool_skill_evolve_rollback, tool_skill_evolve_update,
+    tool_skill_evolve_write_file, tool_skill_read_file,
+};
 use self::system::{tool_location_get, tool_system_time};
 use self::taint::{
     check_taint_net_fetch, check_taint_outbound_header, check_taint_outbound_text,
@@ -5377,621 +5392,6 @@ async fn tool_docker_exec(
     });
 
     serde_json::to_string_pretty(&response).map_err(|e| format!("Serialize error: {e}"))
-}
-
-// ---------------------------------------------------------------------------
-// Persistent process tools
-// ---------------------------------------------------------------------------
-
-/// Start a long-running process (REPL, server, watcher).
-async fn tool_process_start(
-    input: &serde_json::Value,
-    pm: Option<&crate::process_manager::ProcessManager>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let pm = pm.ok_or("Process manager not available")?;
-    let agent_id = caller_agent_id.unwrap_or("default");
-    let command = input["command"]
-        .as_str()
-        .ok_or("Missing 'command' parameter")?;
-    let args: Vec<String> = input["args"]
-        .as_array()
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str().map(String::from))
-                .collect()
-        })
-        .unwrap_or_default();
-
-    let proc_id = pm.start(agent_id, command, &args).await?;
-    Ok(serde_json::json!({
-        "process_id": proc_id,
-        "status": "started"
-    })
-    .to_string())
-}
-
-/// Read accumulated stdout/stderr from a process (non-blocking drain).
-async fn tool_process_poll(
-    input: &serde_json::Value,
-    pm: Option<&crate::process_manager::ProcessManager>,
-) -> Result<String, String> {
-    let pm = pm.ok_or("Process manager not available")?;
-    let proc_id = input["process_id"]
-        .as_str()
-        .ok_or("Missing 'process_id' parameter")?;
-    let (stdout, stderr) = pm.read(proc_id).await?;
-    Ok(serde_json::json!({
-        "stdout": stdout,
-        "stderr": stderr,
-    })
-    .to_string())
-}
-
-/// Write data to a process's stdin.
-async fn tool_process_write(
-    input: &serde_json::Value,
-    pm: Option<&crate::process_manager::ProcessManager>,
-) -> Result<String, String> {
-    let pm = pm.ok_or("Process manager not available")?;
-    let proc_id = input["process_id"]
-        .as_str()
-        .ok_or("Missing 'process_id' parameter")?;
-    let data = input["data"].as_str().ok_or("Missing 'data' parameter")?;
-    // Always append newline if not present (common expectation for REPLs)
-    let data = if data.ends_with('\n') {
-        data.to_string()
-    } else {
-        format!("{data}\n")
-    };
-    pm.write(proc_id, &data).await?;
-    Ok(r#"{"status": "written"}"#.to_string())
-}
-
-/// Terminate a process.
-async fn tool_process_kill(
-    input: &serde_json::Value,
-    pm: Option<&crate::process_manager::ProcessManager>,
-) -> Result<String, String> {
-    let pm = pm.ok_or("Process manager not available")?;
-    let proc_id = input["process_id"]
-        .as_str()
-        .ok_or("Missing 'process_id' parameter")?;
-    pm.kill(proc_id).await?;
-    Ok(r#"{"status": "killed"}"#.to_string())
-}
-
-/// List processes for the current agent.
-async fn tool_process_list(
-    pm: Option<&crate::process_manager::ProcessManager>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let pm = pm.ok_or("Process manager not available")?;
-    let agent_id = caller_agent_id.unwrap_or("default");
-    let procs = pm.list(agent_id);
-    let list: Vec<serde_json::Value> = procs
-        .iter()
-        .map(|p| {
-            serde_json::json!({
-                "id": p.id,
-                "command": p.command,
-                "alive": p.alive,
-                "uptime_secs": p.uptime_secs,
-            })
-        })
-        .collect();
-    Ok(serde_json::Value::Array(list).to_string())
-}
-
-// ---------------------------------------------------------------------------
-// Canvas / A2UI tool
-// ---------------------------------------------------------------------------
-
-/// Sanitize HTML for canvas presentation.
-///
-/// SECURITY: Strips dangerous elements and attributes to prevent XSS:
-/// - Rejects <script>, <iframe>, <object>, <embed>, <applet> tags
-/// - Strips all on* event attributes (onclick, onload, onerror, etc.)
-/// - Strips javascript:, data:text/html, vbscript: URLs
-/// - Enforces size limit
-pub fn sanitize_canvas_html(html: &str, max_bytes: usize) -> Result<String, String> {
-    if html.is_empty() {
-        return Err("Empty HTML content".to_string());
-    }
-    if html.len() > max_bytes {
-        return Err(format!(
-            "HTML too large: {} bytes (max {})",
-            html.len(),
-            max_bytes
-        ));
-    }
-
-    let lower = html.to_lowercase();
-
-    // Reject dangerous tags
-    let dangerous_tags = [
-        "<script", "</script", "<iframe", "</iframe", "<object", "</object", "<embed", "<applet",
-        "</applet",
-    ];
-    for tag in &dangerous_tags {
-        if lower.contains(tag) {
-            return Err(format!("Forbidden HTML tag detected: {tag}"));
-        }
-    }
-
-    // Reject event handler attributes (on*)
-    // Match patterns like: onclick=, onload=, onerror=, onmouseover=, etc.
-    static EVENT_PATTERN: std::sync::LazyLock<regex_lite::Regex> =
-        std::sync::LazyLock::new(|| regex_lite::Regex::new(r"(?i)\bon[a-z]+\s*=").unwrap());
-    if EVENT_PATTERN.is_match(html) {
-        return Err(
-            "Forbidden event handler attribute detected (on* attributes are not allowed)"
-                .to_string(),
-        );
-    }
-
-    // Reject dangerous URL schemes
-    let dangerous_schemes = ["javascript:", "vbscript:", "data:text/html"];
-    for scheme in &dangerous_schemes {
-        if lower.contains(scheme) {
-            return Err(format!("Forbidden URL scheme detected: {scheme}"));
-        }
-    }
-
-    Ok(html.to_string())
-}
-
-// ---------------------------------------------------------------------------
-// Skill evolution tools
-// ---------------------------------------------------------------------------
-
-/// Build the author tag for an agent-triggered evolution. Use the
-/// agent's id so the dashboard history can attribute the change.
-fn agent_author_tag(caller: Option<&str>) -> String {
-    caller
-        .map(|id| format!("agent:{id}"))
-        .unwrap_or_else(|| "agent".to_string())
-}
-
-/// Reject evolution ops when the registry is frozen (Stable mode).
-///
-/// The registry's frozen flag is meant to express "no skill changes in
-/// this kernel", but the evolution module writes to disk directly and
-/// then triggers `reload_skills`, which no-ops under freeze. Without
-/// this gate, an agent running under Stable mode would silently
-/// persist skill mutations that'd be picked up at the next unfreeze
-/// or restart — defeating the whole point of the mode.
-fn ensure_not_frozen(registry: &SkillRegistry) -> Result<(), String> {
-    if registry.is_frozen() {
-        Err("Skill registry is frozen (Stable mode) — skill evolution is disabled".to_string())
-    } else {
-        Ok(())
-    }
-}
-
-async fn tool_skill_evolve_create(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    ensure_not_frozen(registry)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-    let description = input["description"]
-        .as_str()
-        .ok_or("Missing 'description' parameter")?;
-    let prompt_context = input["prompt_context"]
-        .as_str()
-        .ok_or("Missing 'prompt_context' parameter")?;
-    let tags: Vec<String> = input["tags"]
-        .as_array()
-        .map(|a| {
-            a.iter()
-                .filter_map(|v| v.as_str().map(String::from))
-                .collect()
-        })
-        .unwrap_or_default();
-
-    let author = agent_author_tag(caller_agent_id);
-    let skills_dir = registry.skills_dir();
-    match librefang_skills::evolution::create_skill(
-        skills_dir,
-        name,
-        description,
-        prompt_context,
-        tags,
-        Some(&author),
-    ) {
-        Ok(result) => serde_json::to_string(&result).map_err(|e| e.to_string()),
-        Err(e) => Err(format!("Failed to create skill: {e}")),
-    }
-}
-
-async fn tool_skill_evolve_update(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    ensure_not_frozen(registry)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-    let prompt_context = input["prompt_context"]
-        .as_str()
-        .ok_or("Missing 'prompt_context' parameter")?;
-    let changelog = input["changelog"]
-        .as_str()
-        .ok_or("Missing 'changelog' parameter")?;
-
-    // Registry hot-reload happens AFTER the turn finishes, so within
-    // the same turn `create` followed by `update` would find the
-    // registry cache still stale. Fall back to loading straight from
-    // disk when the cache misses — if the skill truly doesn't exist
-    // the helper returns NotFound too.
-    let skill_owned;
-    let skill = match registry.get(name) {
-        Some(s) => s,
-        None => {
-            skill_owned = librefang_skills::evolution::load_installed_skill_from_disk(
-                registry.skills_dir(),
-                name,
-            )
-            .map_err(|e| format!("Skill '{name}' not found: {e}"))?;
-            &skill_owned
-        }
-    };
-
-    let author = agent_author_tag(caller_agent_id);
-    match librefang_skills::evolution::update_skill(skill, prompt_context, changelog, Some(&author))
-    {
-        Ok(result) => serde_json::to_string(&result).map_err(|e| e.to_string()),
-        Err(e) => Err(format!("Failed to update skill: {e}")),
-    }
-}
-
-async fn tool_skill_evolve_patch(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    ensure_not_frozen(registry)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-    let old_string = input["old_string"]
-        .as_str()
-        .ok_or("Missing 'old_string' parameter")?;
-    let new_string = input["new_string"]
-        .as_str()
-        .ok_or("Missing 'new_string' parameter")?;
-    let changelog = input["changelog"]
-        .as_str()
-        .ok_or("Missing 'changelog' parameter")?;
-    let replace_all = input["replace_all"].as_bool().unwrap_or(false);
-
-    // Same-turn create→patch fallback (see tool_skill_evolve_update).
-    let skill_owned;
-    let skill = match registry.get(name) {
-        Some(s) => s,
-        None => {
-            skill_owned = librefang_skills::evolution::load_installed_skill_from_disk(
-                registry.skills_dir(),
-                name,
-            )
-            .map_err(|e| format!("Skill '{name}' not found: {e}"))?;
-            &skill_owned
-        }
-    };
-
-    let author = agent_author_tag(caller_agent_id);
-    match librefang_skills::evolution::patch_skill(
-        skill,
-        old_string,
-        new_string,
-        changelog,
-        replace_all,
-        Some(&author),
-    ) {
-        Ok(result) => serde_json::to_string(&result).map_err(|e| e.to_string()),
-        Err(e) => Err(format!("Failed to patch skill: {e}")),
-    }
-}
-
-async fn tool_skill_evolve_delete(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    ensure_not_frozen(registry)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-
-    // Resolve the actual installed skill's parent directory instead of
-    // blindly targeting `registry.skills_dir() + name`. Workspace skills
-    // shadow global skills with the same name in an agent run; without
-    // this, `skill_evolve_delete` removed the global skill (or reported
-    // NotFound) while leaving the workspace copy the agent was actually
-    // using in place — destructive against the wrong resource.
-    let parent = match registry.get(name) {
-        Some(s) => s
-            .path
-            .parent()
-            .map(|p| p.to_path_buf())
-            .unwrap_or_else(|| registry.skills_dir().to_path_buf()),
-        // Fall back to the global dir when the registry hasn't caught up
-        // yet (e.g. a skill created in this same turn hasn't been
-        // hot-reloaded into the live view) — delete_skill will return
-        // NotFound if nothing exists there either.
-        None => registry.skills_dir().to_path_buf(),
-    };
-    match librefang_skills::evolution::delete_skill(&parent, name) {
-        Ok(result) => serde_json::to_string(&result).map_err(|e| e.to_string()),
-        Err(e) => Err(format!("Failed to delete skill: {e}")),
-    }
-}
-
-async fn tool_skill_evolve_rollback(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    ensure_not_frozen(registry)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-
-    // Same-turn create→rollback fallback (see tool_skill_evolve_update).
-    let skill_owned;
-    let skill = match registry.get(name) {
-        Some(s) => s,
-        None => {
-            skill_owned = librefang_skills::evolution::load_installed_skill_from_disk(
-                registry.skills_dir(),
-                name,
-            )
-            .map_err(|e| format!("Skill '{name}' not found: {e}"))?;
-            &skill_owned
-        }
-    };
-
-    let author = agent_author_tag(caller_agent_id);
-    match librefang_skills::evolution::rollback_skill(skill, Some(&author)) {
-        Ok(result) => serde_json::to_string(&result).map_err(|e| e.to_string()),
-        Err(e) => Err(format!("Failed to rollback skill: {e}")),
-    }
-}
-
-async fn tool_skill_evolve_write_file(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    ensure_not_frozen(registry)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-    let path = input["path"].as_str().ok_or("Missing 'path' parameter")?;
-    let content = input["content"]
-        .as_str()
-        .ok_or("Missing 'content' parameter")?;
-
-    // Same-turn create→write_file fallback.
-    let skill_owned;
-    let skill = match registry.get(name) {
-        Some(s) => s,
-        None => {
-            skill_owned = librefang_skills::evolution::load_installed_skill_from_disk(
-                registry.skills_dir(),
-                name,
-            )
-            .map_err(|e| format!("Skill '{name}' not found: {e}"))?;
-            &skill_owned
-        }
-    };
-
-    match librefang_skills::evolution::write_supporting_file(skill, path, content) {
-        Ok(result) => serde_json::to_string(&result).map_err(|e| e.to_string()),
-        Err(e) => Err(format!("Failed to write file: {e}")),
-    }
-}
-
-async fn tool_skill_evolve_remove_file(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    ensure_not_frozen(registry)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-    let path = input["path"].as_str().ok_or("Missing 'path' parameter")?;
-
-    // Same-turn fallback (see tool_skill_evolve_update).
-    let skill_owned;
-    let skill = match registry.get(name) {
-        Some(s) => s,
-        None => {
-            skill_owned = librefang_skills::evolution::load_installed_skill_from_disk(
-                registry.skills_dir(),
-                name,
-            )
-            .map_err(|e| format!("Skill '{name}' not found: {e}"))?;
-            &skill_owned
-        }
-    };
-
-    match librefang_skills::evolution::remove_supporting_file(skill, path) {
-        Ok(result) => serde_json::to_string(&result).map_err(|e| e.to_string()),
-        Err(e) => Err(format!("Failed to remove file: {e}")),
-    }
-}
-
-/// Read a companion file from an installed skill directory.
-///
-/// Security: resolves the path relative to the skill's installed directory and
-/// rejects any path that escapes via `..` or absolute components. Symlinks are
-/// resolved by `canonicalize()` before the containment check, so a symlink
-/// pointing outside the skill directory is correctly rejected.
-async fn tool_skill_read_file(
-    input: &serde_json::Value,
-    skill_registry: Option<&SkillRegistry>,
-    allowed_skills: Option<&[String]>,
-) -> Result<String, String> {
-    let registry = skill_registry.ok_or("Skill registry not available")?;
-    let skill_name = input["skill"].as_str().ok_or("Missing 'skill' parameter")?;
-    let rel_path = input["path"].as_str().ok_or("Missing 'path' parameter")?;
-
-    // Enforce agent skill allowlist: if the agent specifies allowed skills
-    // (non-empty list), only those skills can be read. Empty = all allowed.
-    if let Some(allowed) = allowed_skills {
-        if !allowed.is_empty() && !allowed.iter().any(|s| s == skill_name) {
-            return Err(format!(
-                "Access denied: agent is not allowed to access skill '{skill_name}'"
-            ));
-        }
-    }
-
-    // Reject absolute paths early — Path::join replaces the base when given
-    // an absolute path, which would bypass the skill directory containment.
-    if std::path::Path::new(rel_path).is_absolute() {
-        return Err("Access denied: absolute paths are not allowed".to_string());
-    }
-
-    // Look up the skill
-    let skill = registry
-        .get(skill_name)
-        .ok_or_else(|| format!("Skill '{}' not found", skill_name))?;
-
-    // Resolve the path relative to the skill directory
-    let requested = skill.path.join(rel_path);
-    let canonical = requested
-        .canonicalize()
-        .map_err(|e| format!("File not found: {}", e))?;
-    let skill_root = skill
-        .path
-        .canonicalize()
-        .map_err(|e| format!("Skill directory error: {}", e))?;
-
-    // Security: ensure the resolved path is within the skill directory
-    if !canonical.starts_with(&skill_root) {
-        return Err(format!(
-            "Access denied: '{}' is outside the skill directory",
-            rel_path
-        ));
-    }
-
-    // Read the file
-    let content = tokio::fs::read_to_string(&canonical)
-        .await
-        .map_err(|e| format!("Failed to read '{}': {}", rel_path, e))?;
-
-    // Fire-and-forget usage tracking — only count when the agent actually
-    // loads the skill's core prompt content, not every supporting file
-    // read. Reading references/templates/scripts/assets shouldn't inflate
-    // the usage metric. Failures (lock contention, disk error) must not
-    // affect tool execution, so we swallow them.
-    let is_core_prompt = matches!(rel_path, "prompt_context.md" | "SKILL.md" | "skill.md");
-    if is_core_prompt {
-        let skill_dir = skill.path.clone();
-        tokio::task::spawn_blocking(move || {
-            if let Err(e) = librefang_skills::evolution::record_skill_usage(&skill_dir) {
-                tracing::debug!(error = %e, dir = %skill_dir.display(), "record_skill_usage failed");
-            }
-        });
-    }
-
-    // Cap output to avoid flooding the context.
-    // Use floor_char_boundary to avoid panicking on multi-byte UTF-8.
-    const MAX_BYTES: usize = 32_000;
-    if content.len() > MAX_BYTES {
-        let truncate_at = content.floor_char_boundary(MAX_BYTES);
-        Ok(format!(
-            "{}\n\n... (truncated at {} bytes, file is {} bytes total)",
-            &content[..truncate_at],
-            truncate_at,
-            content.len()
-        ))
-    } else {
-        Ok(content)
-    }
-}
-
-/// Canvas presentation tool handler.
-async fn tool_canvas_present(
-    input: &serde_json::Value,
-    workspace_root: Option<&Path>,
-) -> Result<String, String> {
-    let html = input["html"].as_str().ok_or("Missing 'html' parameter")?;
-    let title = input["title"].as_str().unwrap_or("Canvas");
-
-    // Use configured max from task-local (set by agent_loop from KernelConfig), or default 512KB.
-    let max_bytes = CANVAS_MAX_BYTES.try_with(|v| *v).unwrap_or(512 * 1024);
-    let sanitized = sanitize_canvas_html(html, max_bytes)?;
-
-    // Generate canvas ID
-    let canvas_id = uuid::Uuid::new_v4().to_string();
-
-    // Save to workspace output directory
-    let output_dir = if let Some(root) = workspace_root {
-        root.join("output")
-    } else {
-        PathBuf::from("output")
-    };
-    let _ = tokio::fs::create_dir_all(&output_dir).await;
-
-    let timestamp = chrono::Utc::now().format("%Y%m%d_%H%M%S");
-    let filename = format!(
-        "canvas_{timestamp}_{}.html",
-        crate::str_utils::safe_truncate_str(&canvas_id, 8)
-    );
-    let filepath = output_dir.join(&filename);
-
-    // Write the full HTML document
-    let full_html = format!(
-        "<!DOCTYPE html>\n<html>\n<head><meta charset=\"utf-8\"><title>{title}</title></head>\n<body>\n{sanitized}\n</body>\n</html>"
-    );
-    tokio::fs::write(&filepath, &full_html)
-        .await
-        .map_err(|e| format!("Failed to save canvas: {e}"))?;
-
-    let response = serde_json::json!({
-        "canvas_id": canvas_id,
-        "title": title,
-        "saved_to": filepath.to_string_lossy(),
-        "size_bytes": full_html.len(),
-    });
-
-    serde_json::to_string_pretty(&response).map_err(|e| format!("Serialize error: {e}"))
-}
-
-// ---------------------------------------------------------------------------
-// Artifact retrieval tool (#3347)
-// ---------------------------------------------------------------------------
-
-/// Implementation of the `read_artifact` tool.
-///
-/// Reads up to `length` bytes from the artifact identified by `handle`,
-/// starting at `offset`.  Both parameters are optional (defaults: 0 and 4096).
-/// The result is UTF-8 text: binary blobs are lossily decoded.
-async fn tool_read_artifact(
-    input: &serde_json::Value,
-    artifact_dir: &std::path::Path,
-) -> Result<String, String> {
-    let handle = input["handle"]
-        .as_str()
-        .ok_or("Missing required parameter 'handle'")?;
-
-    let offset = input["offset"].as_u64().unwrap_or(0) as usize;
-    let length = input["length"]
-        .as_u64()
-        .unwrap_or(4096)
-        .min(crate::artifact_store::MAX_READ_LENGTH as u64) as usize;
-
-    let bytes = crate::artifact_store::read(handle, offset, length, artifact_dir)?;
-
-    if bytes.is_empty() {
-        return Ok(format!(
-            "[read_artifact: {handle} | offset={offset}] — no more content (past end of artifact)"
-        ));
-    }
-
-    let text = String::from_utf8_lossy(&bytes);
-    Ok(format!(
-        "[read_artifact: {handle} | offset={offset} | {} bytes read]\n{text}",
-        bytes.len()
-    ))
 }
 
 #[cfg(test)]
