+++ crates/librefang-runtime/src/tool_runner/process.rs
@@ -1,0 +1,101 @@
+//! Persistent process tools — start / poll / write / kill / list.
+
+/// Start a long-running process (REPL, server, watcher).
+pub(super) async fn tool_process_start(
+    input: &serde_json::Value,
+    pm: Option<&crate::process_manager::ProcessManager>,
+    caller_agent_id: Option<&str>,
+) -> Result<String, String> {
+    let pm = pm.ok_or("Process manager not available")?;
+    let agent_id = caller_agent_id.unwrap_or("default");
+    let command = input["command"]
+        .as_str()
+        .ok_or("Missing 'command' parameter")?;
+    let args: Vec<String> = input["args"]
+        .as_array()
+        .map(|arr| {
+            arr.iter()
+                .filter_map(|v| v.as_str().map(String::from))
+                .collect()
+        })
+        .unwrap_or_default();
+
+    let proc_id = pm.start(agent_id, command, &args).await?;
+    Ok(serde_json::json!({
+        "process_id": proc_id,
+        "status": "started"
+    })
+    .to_string())
+}
+
+/// Read accumulated stdout/stderr from a process (non-blocking drain).
+pub(super) async fn tool_process_poll(
+    input: &serde_json::Value,
+    pm: Option<&crate::process_manager::ProcessManager>,
+) -> Result<String, String> {
+    let pm = pm.ok_or("Process manager not available")?;
+    let proc_id = input["process_id"]
+        .as_str()
+        .ok_or("Missing 'process_id' parameter")?;
+    let (stdout, stderr) = pm.read(proc_id).await?;
+    Ok(serde_json::json!({
+        "stdout": stdout,
+        "stderr": stderr,
+    })
+    .to_string())
+}
+
+/// Write data to a process's stdin.
+pub(super) async fn tool_process_write(
+    input: &serde_json::Value,
+    pm: Option<&crate::process_manager::ProcessManager>,
+) -> Result<String, String> {
+    let pm = pm.ok_or("Process manager not available")?;
+    let proc_id = input["process_id"]
+        .as_str()
+        .ok_or("Missing 'process_id' parameter")?;
+    let data = input["data"].as_str().ok_or("Missing 'data' parameter")?;
+    // Always append newline if not present (common expectation for REPLs)
+    let data = if data.ends_with('\n') {
+        data.to_string()
+    } else {
+        format!("{data}\n")
+    };
+    pm.write(proc_id, &data).await?;
+    Ok(r#"{"status": "written"}"#.to_string())
+}
+
+/// Terminate a process.
+pub(super) async fn tool_process_kill(
+    input: &serde_json::Value,
+    pm: Option<&crate::process_manager::ProcessManager>,
+) -> Result<String, String> {
+    let pm = pm.ok_or("Process manager not available")?;
+    let proc_id = input["process_id"]
+        .as_str()
+        .ok_or("Missing 'process_id' parameter")?;
+    pm.kill(proc_id).await?;
+    Ok(r#"{"status": "killed"}"#.to_string())
+}
+
+/// List processes for the current agent.
+pub(super) async fn tool_process_list(
+    pm: Option<&crate::process_manager::ProcessManager>,
+    caller_agent_id: Option<&str>,
+) -> Result<String, String> {
+    let pm = pm.ok_or("Process manager not available")?;
+    let agent_id = caller_agent_id.unwrap_or("default");
+    let procs = pm.list(agent_id);
+    let list: Vec<serde_json::Value> = procs
+        .iter()
+        .map(|p| {
+            serde_json::json!({
+                "id": p.id,
+                "command": p.command,
+                "alive": p.alive,
+                "uptime_secs": p.uptime_secs,
+            })
+        })
+        .collect();
+    Ok(serde_json::Value::Array(list).to_string())
+}
